Attempted to allow built libraries container to be accessed from others, no luck
I'm stuck because when mounting the container volume to other containers (as read-only volume), I'm possibly depending on libraries that have not been built yet. I need to all containers to build at the same time though.
I think the answer is to create a builder image for libraries, and use that in all other built images (if they depend on a library). This may mean either A) that the build context needs to become the project root so libraries can be accessed (can't access parent folders currently unless context is higher in folder hierarchy) or B) I need to make the project structure resemble a monorepo (not just the libraries folder). Option A sounds like the best choice because 1) I have some projects that aren't JS-based and 2) because projects can be better isolated (I don't have to selectively copy required monorepo parts into each image, since copying the entire monorepo into a container is bad for image size).
Option C) upload libraries to a private proxy registry to be used with PNPM. That's a bad idea though because then I depend on a registry already being set up, but still probably worth considering..