Added Electron as platform for desktop (Tauri is preferable but not ready yet)
Also Tauri platform plugin for Capacitor doesn't have nearly as much activity as the Electron plugin for Capacitor does.

Note that I'll have to use NPM for Electron as it appears there are issues installing with PNPM. I've set the packageManager property for the Electron project and excluded the electron folder from the overall project workspace.