Small changes to docs, possible auto-generated API docs
The auto-generated API docs feels like an info dump and may not be relevant
to someone just wanting to learn the library. Since the library interface is
simple, it may make more sense to manually generate the API docs for the most
part and rely on generated API docs for long list of things like options or
places where comments/types may frequently become out of sync.