Added handling of multiple promises in result (broken right now, WIP)
Once fixed, the limitation is that a callback must be provided on the function (because presence of callback determines if the callback plugin should
be used and only the callback handler supports promises from the server). Since the client has no way to know if the server returns multiple promises
at time that function is called (so client-side plugin can't be determined), there either needs to be some "upgrade" event (makes sense for HTTP->WS
but less sense in other scenarios possibly) to go from method plugin to mallback plugin without calling the function twice, or only the callback plugin
should be used (no method plugin) when multiple returned promises are needed which is unrealistic to expect.