Disabled multiple promise support for now, can be enabled with experimental flag in options The client doesn't know when the server will return multiple promises until the server receives the result of the function. There either needs to be an upgrade event from the method handler to the callback handler (when used on a network, this is HTTP -> WS) or the method handler will need some form of streaming support. Streaming support will be much easier to support (but still difficult).
Today, this only works with the callback handler so a callback must be provided on a function to support
multiple promises. This is a restriction that should be removed in the future. This feature is currently
disabled and placed behind an experimental flag but may be enabled in the future.