Client now supports optional "OverrideModule" type parameter to support partial modules
This is useful when part of a module is available client-side and a call can be made synchronously.
Now, type definitions can optionally reflect when that module is available and avoid wrapping the return result
in a Promise. The default remains the same, and the Module type parameter goes through transformations to reflect
the transport of function data from a remote location (wrapping it in a Promise and adding support for HTML forms).

Currently, the client needs this type parameter passed explicitly (passing partial module to `.module` option alone
isn't enough but I'm unsure how to work around that).